requirements.txt
flask==3.0.0
python-dotenv==1.0.0
hume[stream]
opencv-python-headless
opencv-contrib-python-headless